# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pegmatite Vault API
  version: 0.1.0
  description: 鉱物標本コレクションを管理するローカルWebアプリケーションAPI
servers:
  - url: http://localhost:8000/api
    description: ローカル開発環境
...
paths:
...
  /minerals:
    get:
      tags: [Minerals]
      summary: 鉱物一覧を取得
      operationId: listMinerals
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          description: 和名、英名、化学式の部分一致検索
          schema: { type: string }
        - name: mineral_class_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [japanese_name, english_name, specimen_count]
            default: japanese_name
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: 鉱物一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MineralPage"
    post:
      tags: [Minerals]
      summary: 鉱物を登録
      operationId: createMineral
      requestBody:
        $ref: "#/components/requestBodies/MineralWrite"
      responses:
        "201":
          description: 登録した鉱物
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mineral"
        "422":
          $ref: "#/components/responses/ValidationError"
  /minerals/{mineral_id}:
    parameters:
      - $ref: "#/components/parameters/MineralId"
    get:
      tags: [Minerals]
      summary: 鉱物詳細を取得
      operationId: getMineral
      responses:
        "200":
          description: 鉱物詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mineral"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Minerals]
      summary: 鉱物を更新
      operationId: updateMineral
      requestBody:
        $ref: "#/components/requestBodies/MineralWrite"
      responses:
        "200":
          description: 更新した鉱物
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mineral"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Minerals]
      summary: 鉱物を削除
      operationId: deleteMineral
      responses:
        "204":
          description: 削除成功
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /minerals/{mineral_id}/specimens:
    get:
      tags: [Minerals]
      summary: 鉱物に関連する標本を取得
      operationId: listMineralSpecimens
      parameters:
        - $ref: "#/components/parameters/MineralId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 関連標本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenPage"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /mineral-classes:
    get:
      tags: [Mineral classes]
      summary: 鉱物分類一覧を取得
      operationId: listMineralClasses
      responses:
        "200":
          description: 鉱物分類一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MineralClass"
    post:
      tags: [Mineral classes]
      summary: 鉱物分類を登録
      operationId: createMineralClass
      requestBody:
        $ref: "#/components/requestBodies/NamedMasterWrite"
      responses:
        "201":
          description: 登録した鉱物分類
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MineralClass"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /mineral-classes/{mineral_class_id}:
    parameters:
      - $ref: "#/components/parameters/MineralClassId"
    get:
      tags: [Mineral classes]
      summary: 鉱物分類を取得
      operationId: getMineralClass
      responses:
        "200":
          description: 鉱物分類
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MineralClass"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Mineral classes]
      summary: 鉱物分類を更新
      operationId: updateMineralClass
      requestBody:
        $ref: "#/components/requestBodies/NamedMasterWrite"
      responses:
        "200":
          description: 更新した鉱物分類
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MineralClass"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Mineral classes]
      summary: 鉱物分類を削除
      operationId: deleteMineralClass
      responses:
        "204":
          description: 削除成功
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /acquisition-methods:
    get:
      tags: [Acquisition methods]
      summary: 入手経路一覧を取得
      operationId: listAcquisitionMethods
      responses:
        "200":
          description: 入手経路一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcquisitionMethod"
    post:
      tags: [Acquisition methods]
      summary: 入手経路を登録
      operationId: createAcquisitionMethod
      requestBody:
        $ref: "#/components/requestBodies/NamedMasterWrite"
      responses:
        "201":
          description: 登録した入手経路
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquisitionMethod"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /acquisition-methods/{acquisition_method_id}:
    parameters:
      - $ref: "#/components/parameters/AcquisitionMethodId"
    get:
      tags: [Acquisition methods]
      summary: 入手経路を取得
      operationId: getAcquisitionMethod
      responses:
        "200":
          description: 入手経路
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquisitionMethod"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Acquisition methods]
      summary: 入手経路を更新
      operationId: updateAcquisitionMethod
      requestBody:
        $ref: "#/components/requestBodies/NamedMasterWrite"
      responses:
        "200":
          description: 更新した入手経路
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquisitionMethod"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Acquisition methods]
      summary: 入手経路を削除
      operationId: deleteAcquisitionMethod
      responses:
        "204":
          description: 削除成功
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /settings/storage/validate:
    post:
      tags: [Settings]
      summary: 画像保存先の変更可否を検証
      operationId: validateImageStorageTarget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string, minLength: 1 }
      responses:
        "200":
          description: 移行前検証結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageMigrationPreview"
        "422":
          $ref: "#/components/responses/ValidationError"
  /settings/storage/migrate:
    post:
      tags: [Settings]
      summary: 画像をコピー・照合して保存先を変更
      description: 全ファイルのSHA-256確認後に切り替える。旧保存先は削除しない。
      operationId: migrateImageStorage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string, minLength: 1, maxLength: 1000 }
      responses:
        "200":
          description: 移行結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageMigrationPreview"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
...
    MineralId:
      name: mineral_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MineralClassId:
      name: mineral_class_id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    AcquisitionMethodId:
      name: acquisition_method_id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  requestBodies:
    MineralWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MineralWrite"
...
    NamedMasterWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NamedMasterWrite"
  responses:
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 現在状態または関連データと競合する
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: 入力値が不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [resource_not_found] }
            message: { type: string }
            details:
              type: [object, array, "null"]
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
...
    SpecimenSummary:
      type: object
      required: [id, specimen_no, specimen_name, favorite, mineral_names, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        specimen_no: { type: integer }
        specimen_name: { type: string }
        locality:
          oneOf:
            - $ref: "#/components/schemas/LocalitySummary"
            - type: "null"
        favorite: { type: boolean }
        mineral_names:
          type: array
          items: { type: string }
        thumbnail_image_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    SpecimenPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SpecimenSummary"
...
    MineralWrite:
      type: object
      required: [japanese_name]
      properties:
        japanese_name: { type: string, minLength: 1, maxLength: 255 }
        english_name: { type: [string, "null"], maxLength: 255 }
        formula: { type: [string, "null"], maxLength: 255 }
        crystal_system: { type: [string, "null"], maxLength: 100 }
        mineral_class_id: { type: [integer, "null"], minimum: 1 }
        description: { type: [string, "null"] }
    MineralSummary:
      type: object
      required: [id, japanese_name]
      properties:
        id: { type: string, format: uuid }
        japanese_name: { type: string }
        english_name: { type: [string, "null"] }
        formula: { type: [string, "null"] }
    Mineral:
      allOf:
        - $ref: "#/components/schemas/MineralSummary"
        - type: object
          required: [specimen_count]
          properties:
            crystal_system: { type: [string, "null"] }
            mineral_class:
              oneOf:
                - $ref: "#/components/schemas/MineralClass"
                - type: "null"
            description: { type: [string, "null"] }
            specimen_count: { type: integer, minimum: 0 }
    MineralPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Mineral"
    MineralClass:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, minimum: 1 }
        name: { type: string }
        description: { type: [string, "null"] }
...
    LocalitySummary:
      type: object
      required: [id, locality_name]
      properties:
        id: { type: string, format: uuid }
        locality_name: { type: string }
        alias_name: { type: [string, "null"] }
...
    NamedMasterWrite:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, "null"] }
    AcquisitionMethod:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, minimum: 1 }
        name: { type: string }
        description: { type: [string, "null"] }
...
    StorageMigrationPreview:
      type: object
      required:
        [source_path, target_path, file_count, total_bytes, free_bytes, same_path, ready, issues]
      properties:
        source_path: { type: string }
        target_path: { type: string }
        file_count: { type: integer, minimum: 0 }
        total_bytes: { type: integer, minimum: 0 }
        free_bytes: { type: integer, minimum: 0 }
        same_path: { type: boolean }
        ready: { type: boolean }
        issues:
          type: array
          items: { type: string }
...